character customisation like buttons that change color when they matter

+++ code/modules/interface/keybind_menu_chui.dm
@@ -2,6 +2,8 @@
 /datum/keybind_menu
 	var/client/owner
 	var/list/changed_keys //so we can keep track of what keys the user changes then merge later
+	var/resetting = FALSE
+	var/hasChanges = FALSE
 
 	New(client/my_client)
 		..()
@@ -29,6 +31,8 @@
 			)))
 
 		.["keys"] = keys
+		.["resetting"] = resetting
+		.["hasChanges"] = hasChanges
 
 	ui_act(action, list/params, datum/tgui/ui, datum/ui_state/state)
 		. = ..()
@@ -39,6 +43,7 @@
 			if ("changed_key")
 				if (params["id"] && params["value"])
 					add_to_changed(params["id"], uppertext(params["value"]) )
+					hasChanges = TRUE
 					. = TRUE
 
 			if ("confirm")
@@ -53,14 +58,18 @@
 					current_keymap.on_update(owner)
 					owner.player.cloudSaves.putData("custom_keybind_data", json_encode(changed_keys_rev))
 					boutput(owner, SPAN_NOTICE("Your custom keybinding data has been saved."))
+					hasChanges = FALSE
 					. = TRUE
 			if ("reset")
-				changed_keys = new/list()
-				owner.player.cloudSaves.deleteData("custom_keybind_data")
-				owner.keymap = null //To prevent merge() from not overwriting old keybinds
-				owner.mob.reset_keymap() //Does successive calls to rebuild the keymap
-				boutput(owner, SPAN_NOTICE("Your keybinding data has been reset."))
-				tgui_process.close_uis(src)
+				if (!resetting)
+					resetting = TRUE
+				else
+					changed_keys = new/list()
+					owner.player.cloudSaves.deleteData("custom_keybind_data")
+					owner.keymap = null //To prevent merge() from not overwriting old keybinds
+					owner.mob.reset_keymap() //Does successive calls to rebuild the keymap
+					boutput(owner, SPAN_NOTICE("Your keybinding data has been reset."))
+					tgui_process.close_uis(src)
 			if ("cancel")
 				tgui_process.close_uis(src)
 
@@ -78,4 +87,6 @@
 	if(!src.keybind_menu)
 		src.keybind_menu = new(src)
 	src.keybind_menu.changed_keys = list()
+	src.keybind_menu.resetting = FALSE
+	src.keybind_menu.hasChanges = FALSE
 	src.keybind_menu.ui_interact(src.mob)
